szkiele aplikacji mobilnej

+++ FoodSearch.Data.Mapping/StoredProcedure/Results/RestaurantInfo.cs
@@ -9,7 +9,7 @@ namespace FoodSearch.Data.Mapping.StoredProcedure.Results
         public virtual int LogoId { get; set; }
         public virtual string Street { get; set; }
         public virtual string Number { get; set; }
-        public virtual int MinimumOrder { get; set; }
+        public virtual float MinimumOrder { get; set; }
         public virtual string TimeFrom { get; set; }
         public virtual string TimeTo { get; set; }
         public virtual float RestaurantRating { get; set; }
